deployment/requirements.txt
fastapi
opencv-python-headless
Pillow
pydantic
requests
torch
transformers
uvicorn
python-multipart